Move preview server into storybook middleware
